# fix memory ordering in append_vec