#### Problem
The current memory ordering in AppendVec is not thread-safe. In Relaxed ordering, the compiler could reorder *current_len* write before *map* write in the writer thread.  In such case, when reader tries to access the last element in the map specified by *current_len*, the last element in the map may not have been updated.

#### Summary of Changes
Change all loads of *current_len* to *Acquire* and stores of *current_len* to *Release*. This will force all the write to *map* happens before write to *current_len*. Thus, eliminate the race condition.

Fixes #