@HaoranYi 2022-02-20
@brooksprumo 

```
before:
running 5 tests
test append_vec_append                 ... bench:         339 ns/iter (+/- 3)
test append_vec_concurrent_append_read ... bench:         310 ns/iter (+/- 12)
test append_vec_concurrent_read_append ... bench:         364 ns/iter (+/- 5)
test append_vec_random_read            ... bench:         285 ns/iter (+/- 2)
test append_vec_sequential_read        ... bench:         310 ns/iter (+/- 4)

after: 
running 5 tests
test append_vec_append                 ... bench:         344 ns/iter (+/- 6)
test append_vec_concurrent_append_read ... bench:         317 ns/iter (+/- 2)
test append_vec_concurrent_read_append ... bench:         370 ns/iter (+/- 2)
test append_vec_random_read            ... bench:         292 ns/iter (+/- 2)
test append_vec_sequential_read        ... bench:         314 ns/iter (+/- 10)
```
Benchmark on my own dev machine: AMD Ryzen 7 1700 Eight-Core Processor 3.00 GHz.
The old version relaxed is slightly faster than the new implementation with acq/release, which is expected. 

On x86, the hardware will not reorder the stores. But the optimization compiler can reorder the `self.append_ptr(offset, val.0, val.1)` and `self.current_len.store(*offset, Ordering::Relaxed)`. 

https://github.com/solana-labs/solana/blob/42bdf1d86470d49018f948a03963251c13dfbe07/runtime/src/append_vec.rs#L419 

From the compiler's point of view, there is no dependency in the compiler to force that *append_ptr* happens before *current_len.store*. For Relaxed ordering, compiler may reorder these writes. If that's the case, for example, the writer could update the len to 12 from 10 before it actually writes element 11, 12. Then the reader, could potentially read element 11 and 12 before the writer writes them, which will load bad values. 

With Release and Acquired ordering, the compiler will make sure all writes before are done before current write, and all read before are done before current read. Hence, the case above will not happen.