@brooksprumo 2022-02-21
> From the compiler's point of view, there is no dependency to force that append_ptr happens before current_len.store. For Relaxed ordering, compiler may reorder these writes. If that's the case, the writer could update the len to 12 from 10 before it actually writes element 11, 12. Then the reader, could potentially read element 11 and 12 before the writer writes them, which will load bad values.

@HaoranYi Ah yes, good catch! And thank you for running the benchmarks. Lgtm