@brooksprumo 2022-02-21
@mvines @willhickey @t-nelson What are you thoughts on backporting this PR? It feels like a bugfix to me. v1.8? v1.9? None?