@mvines 2022-02-21
Definitely not v1.8 as this hasn't been observed on mb. 
v1.9 if you feel there's little to no risk, otherwise v1.10 only.