@mvines 2022-02-21
sgtm. we don't actually know that the compiler does the wrong thing here, even though from a theoretical standpoint it might