@brooksprumo 2022-02-21
> sgtm. we don't actually know that the compiler does the wrong thing here, even though from a theoretical standpoint it might

It's not just the compiler, the CPU can reorder these as well. (But x86 is not weakly ordered, so the atomics _likely_ are treated as Acquire/Release. If we wanted to support other architectures (like ARM) that are weaker than x86, then we'd need to worry about this more.)